trim movie step fix
